move the nav bar to a separate component
