+++ src/routes/index.jsx
@@ -1,5 +1,6 @@
 import { A } from "solid-start";
 import Counter from "~/components/Counter";
+
 export default function Home() {
   return (
     <main class="text-center mx-auto text-gray-700 p-4">
